+++ src/main/scala/menger/objects/Builder.scala
@@ -4,6 +4,7 @@ import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.GL20
 import com.badlogic.gdx.graphics.VertexAttributes.Usage
 import com.badlogic.gdx.graphics.g3d.Material
+import com.badlogic.gdx.graphics.g3d.attributes.BlendingAttribute
 import com.badlogic.gdx.graphics.g3d.attributes.ColorAttribute
 import com.badlogic.gdx.graphics.g3d.attributes.IntAttribute
 import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder
@@ -14,12 +15,16 @@ object Builder:
   final val DEFAULT_FLAGS = Usage.Position | Usage.Normal
 
   def material(ambientColor: Color, diffuseColor: Color, specularColor: Color): Material =
-    Material(
+    val mat = Material(
       ColorAttribute.createAmbient(ambientColor),
       ColorAttribute.createDiffuse(diffuseColor),
       ColorAttribute.createSpecular(specularColor),
       IntAttribute.createCullFace(GL20.GL_NONE)
     )
+    // Enable blending if any color has alpha < 1.0
+    if ambientColor.a < 1.0f || diffuseColor.a < 1.0f || specularColor.a < 1.0f then
+      mat.set(new BlendingAttribute(GL20.GL_SRC_ALPHA, GL20.GL_ONE_MINUS_SRC_ALPHA))
+    mat
 
   def material(color: Color): Material =
     material(